backend/requirements.txt
flask 
flask_cors
flask_session
flask_talisman
fitparse
pandas
requests
gunicorn
werkzeug
python-dotenv

